refactor: LCFS - split ComplianceReportSummaryService into focused calculator classes (#4086)
